move add_url method to helpers
